add custom widget supporting memes template (#54)
This adds in a custom widget that was lying around in my personal
fork, and got used to support a Grist template:

  https://templates.getgrist.com/gtzQwTXkgzFG/Meme-Generator

It isn't great code but it works, so just adding it unchanged
other than some comments and formatting. I occasionally zap my
personal fork so it would be safer not to rely on it.