Feature/dot in month view has accent color when selected (#74)
in month view, events that are in boundaries on single day are displayed as a dot. This dot has background color of an event. 
To make event visible as selected on month view, in case of those single day events we need to make dot appear in "selection" color. This Code is to change selected event background to selected color if event is single day and current view is month view, and then change it back to normal background when perspective or selection changed again.