create a widget bundle for insertion into grist-core via npm packaging
This adds a few more flags to the bundling script, and then
uses them in a `prepare` script to assemble a bundle for insertion
into grist-core. The bundle can be previewed as a tarball with:
  yarn pack
or published to the npm registry as @gristlabs/grist-widget with:
  yarn publish --access public

There'll be a companion PR to grist-core to make use of the package.