Save and restore inclusions and exclusions to widget options (#162)
If a user changes which values are included in or excluded from the set
of rows and columns, those changes would not be saved to the widget
options. This means that the next time the widget is loaded, all the
rows and columns are shown again. This commit fixes that by saving the
set of included/excluded rows and columns to the widget options and
restoring them, so that changes to these settings are persisted.