tests: restore gristWebDriverUtils from grist-core
As the header of that file says, it should have never been independently changed. Lots of changes happened to it regardless since the file was copied.

As a first step to fixing the tests, this restores the original copy from grist-core and moves all of the additions to the subclass.